pipeline/requirements.txt
tqdm
torch
open-clip-torch
numpy
fire
requests
ffmpeg-python
yt_dlp
pyarrow
fsspec
webdataset
webvtt-py
wandb
pandas
timeout-decorator
scenedetect
decord
langdetect
torchdata
torchaudio
soundfile
omegaconf
einops
transformers==4.30.0
Pillow
accelerate
bitsandbytes
scipy
imagehash
opencv-python
video2numpy
pydub